requirements.txt
charset-normalizer==3.1.0
Flask==2.3.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
numpy==1.24.3
pandas==2.0.2
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.3.4
gunicorn==20.1.0
google-cloud-storage==2.9.0
protobuf==4.23.2
zipp==3.15.0